# Fix and optimize BasicTemplateSystem