A bug fix and optimizations for BasicTemplateSystem. Each commit message has more details about the changes.