@lspgn 2023-10-30
Thank you very much for this PR! This is a good optimization.

JFYI: I use squash merge, but will try to keep (+reformat) the commits description.

Regarding commit 96a9ad781a7f1c767ed272de62884e2ed6a48407, I haven't tried doing it this way before. I don't think the line:

https://github.com/netsampler/goflow2/blob/96a9ad781a7f1c767ed272de62884e2ed6a48407/decoders/netflow/templates.go#L97

is necessary?
Because it's one allocation, I may not use this commit just yet since I have to make sure calls actually lock properly.

Have you tried benchmarking the two implementations? (I have to write some tests also).

Throwing it out there: with a single level of map, have you thought about [sync.Map](https://pkg.go.dev/sync#Map)?