@lspgn 2023-11-04
Benchmarks from https://github.com/netsampler/goflow2/pull/232 (lower ns/op is better):

Before:
```
BenchmarkTemplatesAdd-4      	 6466957	       160.6 ns/op
BenchmarkTemplatesAddGet-4   	17241684	        68.73 ns/op
```

After:
```
BenchmarkTemplatesAdd-4      	10849370	       112.0 ns/op
BenchmarkTemplatesAddGet-4   	34369524	        33.68 ns/op
```