@mscastanho 2023-11-04
@lspgn sorry for the delay.

Looks like you already have a benchmark, so that's solved.

Regarding avoiding the extra allocation, it's OK to leave it out. The major performance gain from this PR comes from restructuring the map.

And I haven't tested with sync.Map, but it seems to be a good fit for this case. But I'm not familiar with it enough to say if it would make a big difference here.

And thanks for accepting the PR =)